Move schemas to project root
