+++ lib/index.js
@@ -9,7 +9,7 @@ import Joi from "joi";
 
 import compile from "./compile";
 import { entries } from "./util/object";
-import constructorInputSchema from "./schemas/interlock-constructor-input.js";
+import constructorInputSchema from "../schemas/lib/index-constructor-input.js";
 import loadAst from "./compile/modules/load-ast";
 import compileModules from "./compile/modules/compile";
